+++ src/lib/CalendarConfig.ts
